activity: Extract new activity_ui module.
