+++ web/tests/user_search.test.js
@@ -2,16 +2,10 @@
 
 const {strict: assert} = require("assert");
 
-const {mock_esm, set_global, zrequire} = require("./lib/namespace");
+const {mock_esm, zrequire} = require("./lib/namespace");
 const {run_test} = require("./lib/test");
 const $ = require("./lib/zjquery");
 const {page_params} = require("./lib/zpage_params");
-
-const _document = {
-    hasFocus() {
-        return true;
-    },
-};
 
 const fake_buddy_list = {
     scroll_container_sel: "#whatever",
@@ -30,9 +24,7 @@ const presence = mock_esm("../src/presence");
 const resize = mock_esm("../src/resize");
 const sidebar_ui = mock_esm("../src/sidebar_ui");
 
-set_global("document", _document);
-
-const activity = zrequire("activity");
+const activity_ui = zrequire("activity_ui");
 const buddy_data = zrequire("buddy_data");
 const muted_users = zrequire("muted_users");
 const people = zrequire("people");
@@ -71,7 +63,7 @@ function test(label, f) {
         people.add_active_user(me);
         people.initialize_current_user(me.user_id);
         muted_users.set_muted_users([]);
-        activity.set_cursor_and_filter();
+        activity_ui.set_cursor_and_filter();
         f({override});
     });
 }
@@ -111,9 +103,9 @@ test("escape_search", ({override}) => {
     override(popovers, "hide_all", () => {});
 
     set_input_val("somevalue");
-    activity.escape_search();
+    activity_ui.escape_search();
     assert.equal($(".user-list-filter").val(), "");
-    activity.escape_search();
+    activity_ui.escape_search();
     assert.ok($("#user_search_section").hasClass("notdisplayed"));
 
     // We need to reset this because the unit tests aren't isolated from each other.
@@ -132,7 +124,7 @@ test("blur search right", ({override}) => {
 
     $(".user-list-filter").trigger("blur");
     assert.equal($(".user-list-filter").is_focused(), false);
-    activity.initiate_search();
+    activity_ui.initiate_search();
     assert.equal($(".user-list-filter").is_focused(), true);
 });
 
@@ -148,7 +140,7 @@ test("blur search left", ({override}) => {
 
     $(".user-list-filter").trigger("blur");
     assert.equal($(".user-list-filter").is_focused(), false);
-    activity.initiate_search();
+    activity_ui.initiate_search();
     assert.equal($(".user-list-filter").is_focused(), true);
 });
 
@@ -165,7 +157,7 @@ test("filter_user_ids", ({override}) => {
     function test_filter(search_text, expected_users) {
         const expected_user_ids = expected_users.map((user) => user.user_id);
         $(".user-list-filter").val(search_text);
-        const filter_text = activity.get_filter_text();
+        const filter_text = activity_ui.get_filter_text();
         assert.deepEqual(
             buddy_data.get_filtered_and_sorted_user_ids(filter_text),
             expected_user_ids,
@@ -175,7 +167,7 @@ test("filter_user_ids", ({override}) => {
             assert.deepEqual(user_ids, expected_user_ids);
         });
 
-        activity.build_user_sidebar();
+        activity_ui.build_user_sidebar();
     }
 
     // Sanity check data setup.
@@ -233,9 +225,9 @@ test("click on user header to toggle display", ({override}) => {
 });
 
 test("searching", () => {
-    assert.equal(activity.searching(), false);
+    assert.equal(activity_ui.searching(), false);
     $(".user-list-filter").trigger("focus");
-    assert.equal(activity.searching(), true);
+    assert.equal(activity_ui.searching(), true);
     $(".user-list-filter").trigger("blur");
-    assert.equal(activity.searching(), false);
+    assert.equal(activity_ui.searching(), false);
 });
